changed double to float

+++ db/migrate/20110414030811_add_total_distance_to_rides.rb
@@ -1,6 +1,6 @@
 class AddTotalDistanceToRides < ActiveRecord::Migration
   def self.up
-    add_column :rides, :total_distance, :double
+    add_column :rides, :total_distance, :float
   end
 
   def self.down
